refactor(dependency): move `pep621` to `pep621.base`
